Update NetworkLayer.swift
Defined new method to download the image. This method gives the image data from which it is easy to create the image object.

In the earlier method i.e., downloading image as file addition step was required to create the image object from the given url. The process of creating image from the file was taking time.