Update ThingDetailsView.swift
Fixed the Auto layout issue for stack view and image.

Image view had an additional autolayout setting for trailing anchor which was not required. Because the image view had constraint set for top , left, height and width anchors.

The stack view was set to give equal space to its views which was violated by the spacing property.